Refactor Family DAO
