+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/patient/dao/register/RegisterDaoFactory.kt
@@ -19,6 +19,7 @@ package org.smartregister.fhircore.engine.data.local.patient.dao.register
 import javax.inject.Inject
 import javax.inject.Singleton
 import org.smartregister.fhircore.engine.appfeature.model.HealthModule
+import org.smartregister.fhircore.engine.data.local.patient.dao.register.family.FamilyRegisterDao
 import org.smartregister.fhircore.engine.domain.repository.RegisterDao
 
 @Singleton
