+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/patient/dao/register/family/FamilyProfileMapper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.engine.data.local.patient.dao.register
+package org.smartregister.fhircore.engine.data.local.patient.dao.register.family
 
 import com.google.android.fhir.logicalId
 import org.hl7.fhir.r4.model.CarePlan
